tests/wpt/tests/largest-contentful-paint/observe-svg-image.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: observe SVG image.</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
  async_test(function (t) {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
    const beforeLoad = performance.now();
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const url = window.location.origin + '/images/green.svg';
        // green.svg is 100 by 50
        const size = 100 * 50;
        checkImage(entry, url, 'image_id', size, beforeLoad);
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
  }, 'Same-origin image is observable.');
</script>

<img src='/images/green.svg' id='image_id'/>
</body>
